LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/video-poster.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <video poster="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=18054171-69e4-41ab-ba9b-4ae6ee526adc&amp;encodingcheck=&Gbreve;"></video>


FAIL Speculative parsing, document.write(): video-poster Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
